requirements.txt
streamlit
pandas
plotly
oracledb
python-dotenv
python-bcb